Refactor(Settings): SettingsManager 안정성 강화 및 로직 개선
설정 관리 클래스(`SettingsManager`)의 안정성과 확장성을 높이기 위해 에러 처리 방식을 개선하고, 템플릿 구현을
헤더-온리 방식으로 변경함. 또한, 설정 불러오기 시 발생하던 논리 버그를 수정함.

- 주요 변경 사항:
  - 에러 처리 방식 개선 (`std::expected` 도입):
    - `SettingsManager`의 모든 DB 접근 함수의 반환 타입을 `std::expected`로 변경하여, 작업의
성공/실패 여부와 오류 메시지를 명확하게 반환하도록 함.
    - `MainWindow`에서 `SettingsManager` 호출 후 반환값을 확인하여, DB 작업 실패 시
`QMessageBox`를 통해 사용자에게 명시적인 오류 피드백을 제공하도록 수정.

  - 템플릿 리팩토링 (헤더-온리):
    - `saveSetting`, `loadSetting` 템플릿 함수의 구현을 `.cpp` 파일에서 `.h` 헤더 파일로
이동함.
    - 이를 통해 명시적 템플릿 인스턴스화가 필요 없어지고, 향후 새로운 데이터 타입을 추가할 때 유연하게 대처할 수 있도록
확장성을 높임.

  - 런타임 및 논리 버그 수정:
    - `std::bad_variant_access`: `initializeSettingsMap`에서 `defaultValue`를 `config.h` 상수로 교체하여,
      `variant`의 타입 추론을 명확하게 하고 타입 불일치 오류를 해결함.
    - `loadSetting` 함수에서 `if`문 뒤 중괄호(`{}`) 누락으로 인해, DB 값을 읽고도 항상 기본값을
반환하던 치명적인 논리 오류를 수정함.

이번 리팩토링을 통해 `SettingsManager`의 안정성과 코드 품질을 크게 향상시키고, 사용자에게 더 신뢰도 높은 피드백을
제공하도록 개선함.

+++ config.h
@@ -47,7 +47,7 @@ namespace config {
             static constexpr double MinAmplitude = 0.0;
             static constexpr double MaxAmplitude = 500.0;
             static constexpr double DefaultAmplitude = 10.0;
-            static constexpr double DefaultPhaseOffset = 0.0;
+            static constexpr int DefaultPhaseOffset = 0;
         };
 
     };
